Rewrite test for non-introspected objects
